requirements.txt
spotipy
click
pyyaml
